final commt

+++ app/Http/Controllers/DealinController.php
@@ -97,7 +97,9 @@ class DealinController extends Controller
     {
         $dealin = Dealin::where('user_id', Auth::user()->id)->where('id', $id)->first();
         $path = $dealin->file_path;
-        Storage::disk('s3')->delete('images/'.$path);
+        if($path != 'default.jpg') {
+            Storage::disk('s3')->delete('images/' . $path);
+        }
         if ($dealin) {
             $dealin->delete();
             return redirect()->route('mine')->with('success', 'Iklan berhasil dihapus !');
